pyproject.toml
[project]
name = "simple-mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "httpx>=0.28.1",
    "mcp>=1.6.0",
    "ollama>=0.1.5",
    "openai>=1.12.0",
    "prompt-toolkit>=3.0.43",
    "pydantic>=2.11.1",
    "pygments>=2.17.2",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
]
